+++ src/main/java/nextstep/configuration/auth/AuthConfig.java
@@ -1,0 +1,21 @@
+package nextstep.configuration.auth;
+
+import nextstep.member.infrastructure.jwt.JwtTokenProvider;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@Configuration
+public class AuthConfig implements WebMvcConfigurer {
+    private JwtTokenProvider jwtTokenProvider;
+
+    public AuthConfig(JwtTokenProvider jwtTokenProvider) {
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
+    @Override
+    public void addArgumentResolvers(List argumentResolvers) {
+        argumentResolvers.add(new AuthenticationPrincipalArgumentResolver(jwtTokenProvider));
+    }
+}
