+++ src/main/java/nextstep/configuration/auth/AuthenticationPrincipal.java
@@ -1,0 +1,11 @@
+package nextstep.configuration.auth;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface AuthenticationPrincipal {
+}
