+++ src/main/java/nextstep/configuration/auth/AuthenticationPrincipalArgumentResolver.java
@@ -1,0 +1,36 @@
+package nextstep.configuration.auth;
+
+import nextstep.configuration.error.AuthenticationException;
+import nextstep.member.infrastructure.jwt.JwtTokenProvider;
+import nextstep.member.controller.dto.LoginMember;
+import org.springframework.core.MethodParameter;
+import org.springframework.web.bind.support.WebDataBinderFactory;
+import org.springframework.web.context.request.NativeWebRequest;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.method.support.ModelAndViewContainer;
+
+public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
+    private JwtTokenProvider jwtTokenProvider;
+
+    public AuthenticationPrincipalArgumentResolver(JwtTokenProvider jwtTokenProvider) {
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
+    @Override
+    public boolean supportsParameter(MethodParameter parameter) {
+        return parameter.hasParameterAnnotation(AuthenticationPrincipal.class);
+    }
+
+    @Override
+    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
+        String authorization = webRequest.getHeader("Authorization");
+        if (!"bearer".equalsIgnoreCase(authorization.split(" ")[0])) {
+            throw new AuthenticationException();
+        }
+        String token = authorization.split(" ")[1];
+
+        String email = jwtTokenProvider.getPrincipal(token);
+
+        return new LoginMember(email);
+    }
+}
