+++ src/main/java/nextstep/configuration/error/AuthenticationException.java
@@ -1,0 +1,8 @@
+package nextstep.configuration.error;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.UNAUTHORIZED)
+public class AuthenticationException extends RuntimeException {
+}
