--- src/main/java/nextstep/favorite/application/dto/FavoriteRequest.java
@@ -1,22 +1,0 @@
-package nextstep.favorite.application.dto;
-
-public class FavoriteRequest {
-    private Long source;
-    private Long target;
-
-    public FavoriteRequest() {
-    }
-
-    public FavoriteRequest(Long source, Long target) {
-        this.source = source;
-        this.target = target;
-    }
-
-    public Long getSource() {
-        return source;
-    }
-
-    public Long getTarget() {
-        return target;
-    }
-}
