--- src/main/java/nextstep/favorite/application/dto/FavoriteResponse.java
@@ -1,7 +1,0 @@
-package nextstep.favorite.application.dto;
-
-/**
- * TODO: StationResponse를 포함하는 클래스로 만듭니다.
- */
-public class FavoriteResponse {
-}
