--- src/main/java/nextstep/favorite/domain/FavoriteRepository.java
@@ -1,6 +1,0 @@
-package nextstep.favorite.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
-}
