+++ src/main/java/nextstep/favorite/domain/entity/Favorite.java
@@ -1,0 +1,14 @@
+package nextstep.favorite.domain.entity;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class Favorite {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+}
