--- src/main/java/nextstep/member/application/dto/MemberRequest.java
@@ -1,34 +1,0 @@
-package nextstep.member.application.dto;
-
-import nextstep.member.domain.Member;
-
-public class MemberRequest {
-    private String email;
-    private String password;
-    private Integer age;
-
-    public MemberRequest() {
-    }
-
-    public MemberRequest(String email, String password, Integer age) {
-        this.email = email;
-        this.password = password;
-        this.age = age;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
-
-    public Member toMember() {
-        return new Member(email, password, age);
-    }
-}
