--- src/main/java/nextstep/member/application/dto/TokenRequest.java
@@ -1,22 +1,0 @@
-package nextstep.member.application.dto;
-
-public class TokenRequest {
-    private String email;
-    private String password;
-
-    public TokenRequest() {
-    }
-
-    public TokenRequest(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
