+++ src/main/java/nextstep/member/controller/dto/LoginMember.java
@@ -1,0 +1,13 @@
+package nextstep.member.controller.dto;
+
+public class LoginMember {
+    private String email;
+
+    public LoginMember(String email) {
+        this.email = email;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+}
