+++ src/main/java/nextstep/member/controller/dto/MemberResponse.java
@@ -1,0 +1,34 @@
+package nextstep.member.controller.dto;
+
+import nextstep.member.domain.entity.Member;
+
+public class MemberResponse {
+    private Long id;
+    private String email;
+    private Integer age;
+
+    public MemberResponse() {
+    }
+
+    public MemberResponse(Long id, String email, Integer age) {
+        this.id = id;
+        this.email = email;
+        this.age = age;
+    }
+
+    public static MemberResponse of(Member member) {
+        return new MemberResponse(member.getId(), member.getEmail(), member.getAge());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public Integer getAge() {
+        return age;
+    }
+}
