+++ src/main/java/nextstep/member/controller/dto/TokenRequest.java
@@ -1,0 +1,22 @@
+package nextstep.member.controller.dto;
+
+public class TokenRequest {
+    private String email;
+    private String password;
+
+    public TokenRequest() {
+    }
+
+    public TokenRequest(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
