+++ src/main/java/nextstep/member/controller/dto/TokenResponse.java
@@ -1,0 +1,16 @@
+package nextstep.member.controller.dto;
+
+public class TokenResponse {
+    private String accessToken;
+
+    public TokenResponse() {
+    }
+
+    public TokenResponse(String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public String getAccessToken() {
+        return accessToken;
+    }
+}
