--- src/main/java/nextstep/member/domain/LoginMember.java
@@ -1,13 +1,0 @@
-package nextstep.member.domain;
-
-public class LoginMember {
-    private String email;
-
-    public LoginMember(String email) {
-        this.email = email;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-}
