--- src/main/java/nextstep/member/domain/MemberRepository.java
@@ -1,11 +1,0 @@
-package nextstep.member.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
-
-    void deleteByEmail(String email);
-}
