+++ src/main/java/nextstep/member/domain/command/TokenGenerator.java
@@ -1,0 +1,5 @@
+package nextstep.member.domain.command;
+
+public interface TokenGenerator {
+    String createToken(String principal);
+}
