+++ src/main/java/nextstep/member/domain/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package nextstep.member.domain.repository;
+
+import nextstep.member.domain.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByEmail(String email);
+
+    void deleteByEmail(String email);
+}
