+++ src/test/java/nextstep/member/acceptance/AuthAcceptanceTest.java
@@ -3,8 +3,8 @@ package nextstep.member.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.member.domain.Member;
-import nextstep.member.domain.MemberRepository;
+import nextstep.member.domain.entity.Member;
+import nextstep.member.domain.repository.MemberRepository;
 import nextstep.subway.setup.BaseTestSetup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
